auth_api/src/requirements.txt
python-dotenv==0.20.0
psycopg2-binary==2.9.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask==2.1.1
gevent==21.12.0
flask-login==0.6.0
flask-jwt-extended==4.3.1
flask_restful==0.3.9
cryptography==36.0.2
flask-restx==0.5.1
backoff==1.11.1
authlib==1.0.1
requests==2.27.1
grpcio==1.46.0
grpcio-tools==1.46.0
Werkzeug==2.1.2
SQLAlchemy==1.4.9
redis==4.2.2
opentelemetry-api==1.11.1
opentelemetry-sdk==1.11.1
opentelemetry-exporter-jaeger==1.11.1
opentelemetry-instrumentation-flask==0.30b1
pyotp==2.6.0
python-multipart==0.0.5
Flask-xCaptcha==0.5.2
Jinja2==3.0.0
Flask-QRcode==3.0.0